# web/public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Job Tracker API",
    "description": "REST API for the Mini-SaaS Job Application Tracker. Manage job applications, notes, tasks, contacts, and more.\n\n## Authentication\nAll endpoints require authentication via NextAuth JWT session. Include the session cookie in requests.\n\n## Multi-tenant Data Isolation\nAll queries are automatically scoped to the authenticated user. Users can only access their own data.\n\n## Soft Deletes\nRecords are soft-deleted (marked with `deletedAt` timestamp) rather than permanently removed.",
    "version": "1.0.0",
    "contact": {
      "name": "API Support"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/ai/recommendations": {
      "get": {
        "tags": ["AI"],
        "summary": "Get job recommendations (Pro only)",
        "description": "AI-powered job recommendations based on application history",
        "responses": {
          "200": {
            "description": "Job recommendations",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "recommendations": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string" },
                          "company": { "type": "string" },
                          "title": { "type": "string" },
                          "score": { "type": "number" },
                          "reason": { "type": "string" }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/ProRequired" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": { "type": "string" },
              "message": { "type": "string" },
              "details": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "field": { "type": "string" },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "responses": {
      "Unauthorized": {
        "description": "Not authenticated",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": {
              "error": {
                "code": "UNAUTHORIZED",
                "message": "Authentication required"
              }
            }
          }
        }
      },
...
      "ProRequired": {
        "description": "Pro plan required",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": {
              "error": {
                "code": "PRO_REQUIRED",
                "message": "This feature requires a Pro plan"
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "sessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "next-auth.session-token",
        "description": "NextAuth session cookie"
      }
    }
  },
  "security": [
    { "sessionAuth": [] }
  ]
}